pallet_grants::pallet
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_vesting_schedule

Wire funds to be vested by the receiver

bridge_all_vesting_schedules

Initiate a bridge transfer of all vested funds to the given eth_address on the given chain_id. This process will need to be completed by the bridge oracles.

cancel_all_vesting_schedules

Cancel all vested schedules for the given user. If there are coins to be claimed they will be auto claimed for the given user. If limit_to_free_balance is set to true we will not error if the free balance of who has less coins than what was granted and is being revoked (useful if the state was corrupted).

claim

Claim funds that have been vested so far

remove_bridge

Remove a bridge from the list of active bridges.

renounce

Allows the CancelOrigin to renounce to its privileges of being able to cancel who’s vesting schedules.

set_bridge

Allow governance to indicate a bridge is setup between the parachain and a remote chain.