Trait pallet_grants::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self>>;
    type CancelOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type MaxSchedule: Get<u32>;
    type WeightInfo: WeightInfo;
    type BlockNumberProvider: BlockNumberProvider<BlockNumber = BlockNumberFor<Self>>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Currency: LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self>>

type CancelOrigin: EnsureOrigin<Self::RuntimeOrigin>

type MaxSchedule: Get<u32>

The maximum number of vesting schedule.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type BlockNumberProvider: BlockNumberProvider<BlockNumber = BlockNumberFor<Self>>

Object Safety

This trait is not object safe.